feat: 週ごとのPFCカードをホーム画面に移す
レポート画面(/dashboard)への導線がモバイルに無かった。ヘッダーのナビは
`hidden md:flex` でPC幅にしか出ず、下部ナビ(md:hidden)はホーム/体重/食事/運動の
4つでレポートを含まない。つまりスマホからはURL直打ち以外でたどり着けず、
v1.46.0で追加したカードが事実上見えない状態だった。

タブを増やすと320px幅で1タブ64pxまで狭まるため、ナビ構成は触らず、毎日開く
ホーム画面のドットグリッド下にカードを置く。

- useWeeklyPfcTrend: ホーム用に /trends と /user/profile だけを引く軽い hook。
  useDashboard は summary/goals も引くのでホームには重い。
- useDashboard から trendsQuery を削除。レポート画面はもう使わないので、
  残すと毎回無駄なリクエストが1本増える。
